Handle capital return and dividend events better
This fixes the logic with dividends and also adds support for another edge case in both capital return and dividends. Before this, dividends did not support selling an asset and then rebuying it back. It would infact be very wrong if you did that and had number of shares match between the two ownership periods.

Now the calculator supports the case where you sell ALL of an asset and then buy some back again before having asset events for that new ownership period.

Also updated tests to cope with these cases.