Fix merging of same day transactions
Fixes #5.

There was an issue where multiple same-day transactions were merged but then in the resulting output, the merged transaction would get shown along with the non-merged transactions, except the first. This came from the fact that we were merging all transactions into the first on a day but leaving the others lying around.

In this change, I am changing the logic to always create a duplicate transaction for doing the actual processing along with the merging logic. This means that we keep the original transaction objects clean in the input and don't alter them.

As a side effect here I'm also fixing an issue where same day merges would not happen if they are interleaved in the input, i.e. Buy then Sell then Buy for the same asset on the same day.

Tests were updated here, along with adding a new test that would fail before this change.