Support incremental disposals when capital return & dividends exist
Up until now, if you have capital return & dividends (for acc funds) then if there were sales then it must be a sale of all shares at that point. This was to make the calculations simpler. However I needed to account for partial sales. So this diff adds that functionality.

I couldn't find a definitive answer to how to handle these situations, but the general consensus seems to be that you account for disposals in a FIFO order for the purposes of capital return & dividends.

So take this scenario:
1) 01/01/2020: BUY 100 shares
2) 01/04/2020: Capital return event (equalisation part of dividend). Attributed to the buy at 1.
3) 01/01/2021: BUY 100 shares
4) 01/03/2021: SELL 100 shares ("matched" against buy at 1).
5) 01/04/2021: Capital return event (equalisation part of dividend). Attributed to the buy at 3.