Create a `Presenter` protocol
Put `TextPresenter` behind a new protocol called `Presenter`. Then we can have other presenters, like maybe a PDF one.

Also this change allows presenters to return either a string or data. This will be useful for other non-text based presenters to allow them to return pure data. If it returns data then it can't be output to the console.